refactor: move comments and communities to features
