+++ src/components/features/communities/EditCommunityForm.tsx
@@ -13,17 +13,17 @@ import {
   FormMessage,
 } from '@/components/ui/Form';
 import { toast } from '@/hooks/use-toast';
-import { Button } from '../ui/Button';
-import { Input } from '../ui/Input';
+import { Button } from '../../ui/Button';
+import { Input } from '../../ui/Input';
 import { useMutation } from '@tanstack/react-query';
 import axios, { AxiosError } from 'axios';
 import { useRouter } from 'next/navigation';
 import { useCustomToast } from '@/hooks/use-custom-toast';
-import { DropZone } from '../ui/UploadImage';
+import { DropZone } from '../../ui/UploadImage';
 import { startTransition } from 'react';
 import { EditCommunityValidator } from '@/validators/community';
 import { Community } from '@prisma/client';
-import UserAvatar from '../ui/UserAvatar';
+import UserAvatar from '../../ui/UserAvatar';
 
 interface EditCommunityFormProps {
   community: Community;
